SQL Database of International Debt Data
Utilized SQL to create an exceptional database, which is based on data provided by The World Bank between the years 1970 and 2015. This data contains the amount of debt (in USD) owed by developing countries all over the globe, which are all based on several different categories that they all fall in.