USPS COVID-19 Solution for Global Enterprises
Utilizes SQL to develop an open-source tool that allows companies to fully track over hundreds of thousands of packages simultaneously. This solved a huge issue for multiple companies due to the COVID-19 Pandemic. This has promoted USPS customer retention and customer satisfaction, as well as a sales increase during the peak of the COVID-19 Pandemic. 

**IMPORTANT**
This can only be used on Microsoft Access.